Fix build of save-core artifacts
### What's done:
* Refactoring of integration tests
