+++ save-core/src/commonNonJsTest/kotlin/org/cqfn/save/core/test/utils/TestUtilsCommon.kt
@@ -10,14 +10,18 @@ import org.cqfn.save.core.Save
 import org.cqfn.save.core.config.OutputStreamType
 import org.cqfn.save.core.config.ReportType
 import org.cqfn.save.core.config.SaveProperties
+import org.cqfn.save.core.files.createFile
+import org.cqfn.save.core.logging.logDebug
 import org.cqfn.save.core.result.Pass
 import org.cqfn.save.reporter.test.TestReporter
 
 import io.ktor.client.*
+import io.ktor.client.call.receive
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import io.ktor.http.*
 import okio.FileSystem
+import okio.Path.Companion.toPath
 
 import kotlin.test.assertEquals
 import kotlinx.coroutines.CoroutineScope
@@ -34,11 +38,38 @@ const val TIMEOUT = 100_000L
 expect fun runTest(block: suspend (scope: CoroutineScope) -> Unit)
 
 /**
- * @param url
- * @param fileName
- * @return path of downloaded file
+ * Creates an [HttpClient] on a particular platform
+ *
+ * @return a new instance of [HttpClient]
  */
-expect suspend fun downloadFile(url: String, fileName: String): String
+expect suspend fun createHttpClient(): HttpClient
+
+/**
+ * Download file from [url] into [fileName]
+ *
+ * @param url url to download from
+ * @param fileName name of the downloaded file
+ * @return path to the downloaded file as a string
+ */
+suspend fun downloadFile(url: String, fileName: String): String {
+    val fs = FileSystem.SYSTEM
+    val client = createHttpClient()
+
+    val file = fileName.toPath()
+    if (!fs.exists(file)) {
+        val httpResponse: HttpResponse = client.get(url)
+        val responseBody: ByteArray = httpResponse.receive()
+        logDebug("Writing ${responseBody.size} bytes into $file")
+        val newPath = fs.createFile(fileName)
+        fs.write(newPath) {
+            write(responseBody)
+        }
+        logDebug("$url downloaded to $file")
+    }
+
+    client.close()
+    return file.toString()
+}
 
 /**
  * @param testDir
