+++ save-core/src/nativeTest/kotlin/org/cqfn/save/core/test/utils/TestUtilsNative.kt
@@ -6,16 +6,11 @@
 
 package org.cqfn.save.core.test.utils
 
-import org.cqfn.save.core.files.createFile
-import org.cqfn.save.core.logging.logDebug
-
 import io.ktor.client.*
 import io.ktor.client.call.*
 import io.ktor.client.features.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
-import okio.FileSystem
-import okio.Path.Companion.toPath
 
 import kotlinx.coroutines.*
 import kotlinx.coroutines.CoroutineScope
@@ -28,35 +23,11 @@ import kotlinx.coroutines.CoroutineScope
  */
 actual fun runTest(block: suspend (scope: CoroutineScope) -> Unit) = runBlocking { block(this) }
 
-/**
- * @param url
- * @param fileName
- * @return path of the downloaded file
- */
-actual suspend fun downloadFile(url: String, fileName: String): String {
-    val fs = FileSystem.SYSTEM
-
-    val client = HttpClient {
-        install(HttpTimeout) {
-            requestTimeoutMillis = TIMEOUT
-            connectTimeoutMillis = TIMEOUT
-            socketTimeoutMillis = TIMEOUT
-        }
+@Suppress("MISSING_KDOC_ON_FUNCTION", "MISSING_KDOC_TOP_LEVEL")  // https://github.com/cqfn/diKTat/issues/1012
+actual suspend fun createHttpClient() = HttpClient {
+    install(HttpTimeout) {
+        requestTimeoutMillis = TIMEOUT
+        connectTimeoutMillis = TIMEOUT
+        socketTimeoutMillis = TIMEOUT
     }
-
-    val file = fileName.toPath()
-
-    if (!fs.exists(file)) {
-        val httpResponse: HttpResponse = client.get(url)
-        val responseBody: ByteArray = httpResponse.receive()
-        logDebug("Writing ${responseBody.size} bytes into $file")
-        val newPath = fs.createFile(fileName)
-        fs.write(newPath) {
-            write(responseBody)
-        }
-        logDebug("$url downloaded to $file")
-    }
-
-    client.close()
-    return file.toString()
 }
